+++ changelog.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ##### Improvements:
 1. The firmware loading indicator on the chip has been reduced from three integers to two.
 2. Optimized display of links in the log.
+3. For the slave, a button for copying dsk to the clipboard has been added.
 
 ##### Fixes:
 1. Fixed an error that occurred when an open COM port disappeared and resulted in the inability to close this utility.